Extract inner ScheduleDetails component from SchedulePage
This splits the responsibilities up a bit:
- Schedule page: selects a Schedule resource by actor, and loads it
- ScheduleDetails component: given a Schedule resource, loads the
  related Slot/Appointment resources and renders the calendar UI

This division will hopefully help keep things a little tidier; because
the inner component always knows it has a Schedule resource, it can
remove some checks. The outer component manages a much smaller set of
states, and provides a good place for us to hook other enhancements in
to the loading process as a follow up.
