+++ examples/medplum-provider/src/pages/schedule/SchedulePage.module.css
@@ -2,29 +2,3 @@
   background-color: light-dark(var(--mantine-color-white), var(--mantine-color-dark-8));
   height: 100vh;
 }
-
-.wrapper {
-  display: flex;
-  flex-direction: column;
-  height: 100%;
-  width: 100%;
-  overflow-x: auto;
-}
-
-.container {
-  display: flex;
-  flex: 1;
-  min-height: 0;
-  gap: var(--mantine-spacing-sm);
-}
-
-.calendar {
-  flex-grow: 1;
-  min-height: 0;
-  min-width: 0;
-  overflow-x: auto;
-}
-
-.findPane {
-  flex-basis: 25%;
-}
